requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.9
supabase==1.2.0
python-dotenv==1.0.0
requests==2.31.0
flask-cors==4.0.0
httpx==0.24.1
openai==1.3.0
langdetect==1.0.9
Werkzeug==2.3.7

# Production dependencies
Flask-Limiter==3.5.0
gunicorn==21.2.0
prometheus-client==0.19.0
structlog==23.2.0
sentry-sdk[flask]==1.40.0
redis==5.0.1
celery==5.3.4
flask-compress==1.14
flask-talisman==1.1.0 